model access control changes rebased to 2.x (#902)
* Backport changes to model-access-control feature branch (#837)

* rename model meta/chunk API (#827)


* Change the ziputil dependency to fix a potential security concern (#824)


* add text docs ML input (#830)


---------


* add model group (#840)

* add model group


* rename create model group as register model group


* fix class name in build.gradle


* remove unused code; stash thread context


* exclude class for low coverage


---------


* access validation for register/update model-group, register/get/delete/deploy/predict model


* changes to security utils and register/search model group


* update last_updated_time in model group when new version added


* fix undeploy model action


* fix format violations


* rebased to 2.x and prediction/search API fix


* fix rebase conflicts


* new delete model group API


* fix formal violations


---------

+++ plugin/src/main/java/org/opensearch/ml/action/models/DeleteModelTransportAction.java
@@ -9,6 +9,7 @@ import static org.opensearch.common.xcontent.XContentParserUtils.ensureExpectedT
 import static org.opensearch.ml.common.CommonValue.ML_MODEL_INDEX;
 import static org.opensearch.ml.common.MLModel.ALGORITHM_FIELD;
 import static org.opensearch.ml.common.MLModel.MODEL_ID_FIELD;
+import static org.opensearch.ml.settings.MLCommonsSettings.ML_COMMONS_VALIDATE_BACKEND_ROLES;
 import static org.opensearch.ml.utils.MLNodeUtils.createXContentParserFromRegistry;
 import static org.opensearch.ml.utils.RestActionUtils.getFetchSourceContext;
 
@@ -23,8 +24,11 @@ import org.opensearch.action.get.GetResponse;
 import org.opensearch.action.support.ActionFilters;
 import org.opensearch.action.support.HandledTransportAction;
 import org.opensearch.client.Client;
+import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.common.inject.Inject;
+import org.opensearch.common.settings.Settings;
 import org.opensearch.common.util.concurrent.ThreadContext;
+import org.opensearch.commons.authuser.User;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.index.query.TermsQueryBuilder;
@@ -33,10 +37,13 @@ import org.opensearch.index.reindex.DeleteByQueryAction;
 import org.opensearch.index.reindex.DeleteByQueryRequest;
 import org.opensearch.ml.common.MLModel;
 import org.opensearch.ml.common.exception.MLResourceNotFoundException;
+import org.opensearch.ml.common.exception.MLValidationException;
 import org.opensearch.ml.common.model.MLModelState;
 import org.opensearch.ml.common.transport.model.MLModelDeleteAction;
 import org.opensearch.ml.common.transport.model.MLModelDeleteRequest;
 import org.opensearch.ml.common.transport.model.MLModelGetRequest;
+import org.opensearch.ml.utils.RestActionUtils;
+import org.opensearch.ml.utils.SecurityUtils;
 import org.opensearch.rest.RestStatus;
 import org.opensearch.search.fetch.subphase.FetchSourceContext;
 import org.opensearch.tasks.Task;
@@ -49,7 +56,7 @@ import lombok.experimental.FieldDefaults;
 import lombok.extern.log4j.Log4j2;
 
 @Log4j2
-@FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
+@FieldDefaults(level = AccessLevel.PRIVATE)
 public class DeleteModelTransportAction extends HandledTransportAction<ActionRequest, DeleteResponse> {
 
     static final String TIMEOUT_MSG = "Timeout while deleting model of ";
@@ -58,17 +65,25 @@ public class DeleteModelTransportAction extends HandledTransportAction<ActionReq
     static final String OS_STATUS_EXCEPTION_MESSAGE = "Failed to delete all model chunks";
     Client client;
     NamedXContentRegistry xContentRegistry;
+    ClusterService clusterService;
+
+    private volatile boolean filterByEnabled;
 
     @Inject
     public DeleteModelTransportAction(
         TransportService transportService,
         ActionFilters actionFilters,
         Client client,
-        NamedXContentRegistry xContentRegistry
+        NamedXContentRegistry xContentRegistry,
+        Settings settings,
+        ClusterService clusterService
     ) {
         super(MLModelDeleteAction.NAME, transportService, actionFilters, MLModelDeleteRequest::new);
         this.client = client;
         this.xContentRegistry = xContentRegistry;
+        this.clusterService = clusterService;
+        filterByEnabled = ML_COMMONS_VALIDATE_BACKEND_ROLES.get(settings);
+        clusterService.getClusterSettings().addSettingsUpdateConsumer(ML_COMMONS_VALIDATE_BACKEND_ROLES, it -> filterByEnabled = it);
     }
 
     @Override
@@ -78,6 +93,7 @@ public class DeleteModelTransportAction extends HandledTransportAction<ActionReq
         MLModelGetRequest mlModelGetRequest = new MLModelGetRequest(modelId, false);
         FetchSourceContext fetchSourceContext = getFetchSourceContext(mlModelGetRequest.isReturnContent());
         GetRequest getRequest = new GetRequest(ML_MODEL_INDEX).id(modelId).fetchSourceContext(fetchSourceContext);
+        User user = RestActionUtils.getUserContext(client);
 
         try (ThreadContext.StoredContext context = client.threadPool().getThreadContext().stashContext()) {
             client.get(getRequest, ActionListener.wrap(r -> {
@@ -90,37 +106,48 @@ public class DeleteModelTransportAction extends HandledTransportAction<ActionReq
                             algorithmName = getResponse.getSource().get(ALGORITHM_FIELD).toString();
                         }
                         MLModel mlModel = MLModel.parse(parser, algorithmName);
-                        MLModelState mlModelState = mlModel.getModelState();
-                        if (mlModelState.equals(MLModelState.LOADED)
-                            || mlModelState.equals(MLModelState.LOADING)
-                            || mlModelState.equals(MLModelState.PARTIALLY_LOADED)
-                            || mlModelState.equals(MLModelState.DEPLOYED)
-                            || mlModelState.equals(MLModelState.DEPLOYING)
-                            || mlModelState.equals(MLModelState.PARTIALLY_DEPLOYED)) {
-                            actionListener
-                                .onFailure(
-                                    new Exception(
-                                        "Model cannot be deleted in deploying or deployed state. Try undeploy model first then delete"
-                                    )
-                                );
-                        } else {
-                            DeleteRequest deleteRequest = new DeleteRequest(ML_MODEL_INDEX, modelId);
-                            client.delete(deleteRequest, new ActionListener<DeleteResponse>() {
-                                @Override
-                                public void onResponse(DeleteResponse deleteResponse) {
-                                    deleteModelChunks(modelId, deleteResponse, actionListener);
-                                }
 
-                                @Override
-                                public void onFailure(Exception e) {
-                                    log.error("Failed to delete model meta data for model: " + modelId, e);
-                                    if (e instanceof ResourceNotFoundException) {
-                                        deleteModelChunks(modelId, null, actionListener);
-                                    }
-                                    actionListener.onFailure(e);
+                        SecurityUtils.validateModelGroupAccess(user, mlModel.getModelGroupId(), client, ActionListener.wrap(access -> {
+                            if ((filterByEnabled) && (!access)) {
+                                actionListener
+                                    .onFailure(new MLValidationException("User Doesn't have previlege to perform this operation"));
+                            } else {
+                                MLModelState mlModelState = mlModel.getModelState();
+                                if (mlModelState.equals(MLModelState.LOADED)
+                                    || mlModelState.equals(MLModelState.LOADING)
+                                    || mlModelState.equals(MLModelState.PARTIALLY_LOADED)
+                                    || mlModelState.equals(MLModelState.DEPLOYED)
+                                    || mlModelState.equals(MLModelState.DEPLOYING)
+                                    || mlModelState.equals(MLModelState.PARTIALLY_DEPLOYED)) {
+                                    actionListener
+                                        .onFailure(
+                                            new Exception(
+                                                "Model cannot be deleted in deploying or deployed state. Try undeploy model first then delete"
+                                            )
+                                        );
+                                } else {
+                                    DeleteRequest deleteRequest = new DeleteRequest(ML_MODEL_INDEX, modelId);
+                                    client.delete(deleteRequest, new ActionListener<DeleteResponse>() {
+                                        @Override
+                                        public void onResponse(DeleteResponse deleteResponse) {
+                                            deleteModelChunks(modelId, deleteResponse, actionListener);
+                                        }
+
+                                        @Override
+                                        public void onFailure(Exception e) {
+                                            log.error("Failed to delete model meta data for model: " + modelId, e);
+                                            if (e instanceof ResourceNotFoundException) {
+                                                deleteModelChunks(modelId, null, actionListener);
+                                            }
+                                            actionListener.onFailure(e);
+                                        }
+                                    });
                                 }
-                            });
-                        }
+                            }
+                        }, e -> {
+                            log.error("Failed to validate Access for Model Id " + modelId, e);
+                            actionListener.onFailure(e);
+                        }));
                     } catch (Exception e) {
                         log.error("Failed to parse ml model" + r.getId(), e);
                         actionListener.onFailure(e);
@@ -148,7 +175,8 @@ public class DeleteModelTransportAction extends HandledTransportAction<ActionReq
                 if (deleteResponse != null) {
                     // If model metaData not found and deleteResponse is null, do not return here.
                     // ResourceNotFound is returned to notify that this model was deleted.
-                    // This is a walk around to avoid cleaning up model leftovers. Will revisit if necessary.
+                    // This is a walk around to avoid cleaning up model leftovers. Will revisit if
+                    // necessary.
                     actionListener.onResponse(deleteResponse);
                 }
             } else {
